wip: firebase auth

+++ src/components/Main.js
@@ -1,5 +1,6 @@
 import { useEffect, useState } from 'react'
-import { Routes, Route } from "react-router-dom"
+import { Routes, Route, useNavigate } from "react-router-dom"
+
 import Index from "../pages/Index";
 import Show from "../pages/Show";
 import LandingPage from './LandingPage';
@@ -8,6 +9,7 @@ const Main = (props) => {
     
     const [ job, setJob ] = useState(null)
     const url = "http://localhost:4000/jobs/"
+    const navigate = useNavigate()
 
     const getJob = async () => {
         const response = await fetch(url)
@@ -18,6 +20,7 @@ const Main = (props) => {
     
     
     const createJob = async (posting) => {
+        if (!props.user) return
         console.log('posting', posting)
         await fetch(url, { 
             method: "POST",
@@ -32,6 +35,7 @@ const Main = (props) => {
 
 
     const updateJob = async (posting, id) => {
+        if (!props.user) return
         await fetch (url + id, {
             method: "PUT", 
             headers: {
@@ -43,6 +47,7 @@ const Main = (props) => {
     }
 
     const deleteJob = async (id) => {
+        if (!props.user) return
         await fetch(url + id, {
             method: "DELETE",
         })
@@ -56,8 +61,8 @@ const Main = (props) => {
     <main>
     <Routes> 
         <Route path="/" element={<LandingPage />} />
-        <Route exact path="/jobs" element={<Index job={job} createJob={createJob} updateJob={updateJob} deleteJob={deleteJob} /> } />
-        <Route path="/jobs/:id" element={<Show job={job} updateJob={updateJob} deleteJob={deleteJob}/> } /> 
+        <Route exact path="/jobs" element={<Index user={props.user} job={job} createJob={createJob} updateJob={updateJob} deleteJob={deleteJob} /> } />
+        <Route path="/jobs/:id"  element={ props.user ? <Show job={job} updateJob={updateJob} deleteJob={deleteJob}/> : navigate("/")  } /> 
     </Routes>
     </main>
